+++ stdnet/backends/__init__.py
@@ -2,7 +2,3 @@ from .base import *
 from .main import *
 
 
-class BackendRequest(object):
-    pass
-
-