+++ stdnet/lib/redis/lua/zpop.lua
@@ -1,7 +1,32 @@
 -- REDIS ZPOP command. you can pop one or more elements form the sorted set
+-- either by range or by score
 local key = KEYS[1]
-local start = KEYS[2]
-local stop = KEYS[3]
-local range = redis.call('ZRANGE', key, start, stop)
-redis.call('ZREMRANGEBYRANK', key, start, stop)
+local start = ARGV[2]
+local stop = ARGV[3]
+local desc = ARGV[4] + 0
+local args = {start, stop}
+local range
+
+if ARGV[5] + 0 == 1 then
+    args[3] = 'withscores'
+end
+
+-- POP by RANK
+if ARGV[1] == 'rank' then
+    if desc == 1 then
+        range = redis.call('ZREVRANGE', key, unpack(args))
+    else
+        range = redis.call('ZRANGE', key, unpack(args))
+    end
+    redis.call('ZREMRANGEBYRANK', key, start, stop)
+-- POP by SCORE
+else
+    if desc == 1 then
+        range = redis.call('ZrevRANGEBYSCORE', key, unpack(args))
+    else
+        range = redis.call('ZRANGEBYSCORE', key, unpack(args))
+    end
+    redis.call('ZREMRANGEBYSCORE', key, start, stop)
+end
+
 return range